requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==1.5.0
python-dotenv==1.0.1
pillow==10.4.0
python-multipart==0.0.9
httpx==0.28.1
